Add files via upload
This code has been modified from the code from DutchRican below.  I edited a header with pinouts and other minor fixes.  
The commented out code, is uncommented allowing the code to actually run the car.


VM_maker_robot_car
This is the code that was included with a light refactor. Pay attention to where you are actually connecting the drive module as the code was wrong there. Either fix on the board or in the code.

warning!
My board had some issues too, but this is not related to the pdf that was shared from http://www.cs.columbia.edu/~sedwards/presentations/robot-car-instructions.pdf

That is a different layout and the connections are different

Files from DVD
These are files from the DVD that is included

building instructions not by me! This is a nice and helpful person
https://www.youtube.com/watch?v=d9TXThtSUNM